requirements.txt
django-haystack
Mezzanine==3.1.10
django==1.7
django-tastypie
djangorestframework
elasticsearch
